+++ docs/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Development version
+
+### Breaking changes
+
+* The `jarl.toml` argument `assignment` (deprecated since 0.5.0) is removed. Use
+  the rule-specific option `[lint.assignment]` instead (#663).
+
 ## 0.6.0
 
 ::: {.callout-note icon=false title="Released on 2026-08-24" .low-opacity}